# Closes #893: improve 'file prompt quiet' efficiency