I tested this with the following simple snippet of code:

```python
from napalm import get_network_driver


driver = get_network_driver("ios")
dev = driver('172.21.129.3', 'user', 'pass')
dev.open()
dev.load_merge_candidate(config="username xxx secret xxx")
dev.commit_config()
dev.close()
```

Before the change:

```shell
time pipenv run python testing.py
pipenv run python testing.py  1.31s user 0.89s system 2% cpu 1:48.68 total
```

After the change:

```shell
time pipenv run python testing.py
pipenv run python testing.py  1.21s user 0.79s system 2% cpu 1:08.55 total
```

A bigger gain will be obtained especially if other operations happen (e.g if one discards the changes). **At a minimum, this change will speed-up NAPALM (ios) by 40%.**